// src/types/theme.ts
export interface ThemeState {
  colors: {
    white: string;
    primary10: string;
    primary40: string;
    primary90: string;
    secondary10: string;
    secondary40: string;
    secondary90: string;
    tertiary10: string;
    tertiary40: string;
    tertiary90: string;
    neutral10: string;
    neutral99: string;
    neutralVariant30: string;
    neutralVariant50: string;
    neutralVariant80: string;
    neutralVariant90: string;
    surface: string;
    surface1: string;
    surface2: string;
    surface3: string;
    surface4: string;
    surface5: string;
    error10: string;
    error40: string;
    error90: string;
    onSurface12: string;
    onSurface38: string;
  };
  fontSize: {
    display: {
      large: string;
      medium: string;
      small: string;
    };
    headline: {
      large: string;
      medium: string;
      small: string;
    };
    label: {
      large: string;
      medium: string;
      small: string;
    };
    body: {
      large: string;
      medium: string;
      small: string;
    };
  };
}